pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perceptra-detector"
version = "0.1.0"
description = "Unified, production-ready detection and segmentation framework supporting YOLO, DETR, RT-DETR, and custom models."
authors = [{ name = "Tannous Geagea", email = "tannousgeagea@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = [
    "object-detection",
    "segmentation",
    "yolo",
    "detr",
    "rt-detr",
    "computer-vision",
    "deep-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
# Model backends
yolo = [
    "ultralytics>=8.0.0",
]
detr = [
    "transformers>=4.30.0",
]
rfdetr = [
    "rfdetr @ git+https://github.com/roboflow/rf-detr.git",
]
# API server dependencies
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "prometheus-client>=0.19.0",
]
# Developer tooling
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "httpx>=0.25.0",
]
# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]
# Convenience alias for all optional deps
all = [
    "perceptra-detector[yolo,detr,rfdetr,server,dev,docs]",
]

[project.scripts]
# CLI entrypoint
perceptra-detector = "perceptra_detector.cli.commands:main"

[project.urls]
Homepage = "https://github.com/tannousgeagea/perceptra-detector"
Documentation = "https://perceptra-detector.readthedocs.io"
Repository = "https://github.com/tannousgeagea/perceptra-detector"

[tool.setuptools.packages.find]
include = ["perceptra_detector*", "configs*", "examples*", "tests*"]
exclude = ["training/*", "docker*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=perceptra_detector --cov-report=html --cov-report=term"
